refactoring to fix test
