+++ src/services/database.py
@@ -4,12 +4,13 @@ from sqlalchemy.orm import sessionmaker
 from dotenv import load_dotenv
 import os
 
-load_dotenv()
-SQLALCHEMY_DATABASE_URL = os.getenv("DB_CONNECTION_STRING")
+# load_dotenv()
+# SQLALCHEMY_DATABASE_URL = os.getenv("DB_CONNECTION_STRING")
 
-engine = create_engine(
-    SQLALCHEMY_DATABASE_URL
-)
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+# engine = create_engine(
+#     SQLALCHEMY_DATABASE_URL
+# )
 
-Base = declarative_base()
+#SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+# Base = declarative_base()
